Enhance Uploading... progress report
When being used with rake, the output is buffered. The progress report
lines "Uploading xxx" ends up being spurt in blocks when the buffer is
exhausted.

Use STDOUT.flush to explicitly dump the line when output is buffered.

While at it append 'OK' on success, and 'FAILED' whenever an exception
happens.

Closes #11